SQL型の修正 serial→bigserial、int→bigint

+++ ddl/task_ex.sql
@@ -1,5 +1,5 @@
 CREATE TABLE TASK_EX(
-  task_id int primary key references TASK(id) on delete cascade,
+  task_id bigint primary key references TASK(id) on delete cascade,
   due_date timestamp,
   created_at timestamp not null default current_timestamp,
   updated_at timestamp not null default current_timestamp
